Refactored the javascript filtering. Moved the localstorage hook to a new hooks folder. Moved the parseIssueTypesFilter to the index file and renamed it to be more generic since it should work for all filters as long as the database values match the filter values.
