+++ client/app/nonComp/components/TaskTableTab.jsx
@@ -17,7 +17,8 @@ import {
 } from '../../queue/components/TaskTableColumns';
 import {
   buildDecisionReviewFilterInformation,
-  extractEnabledTaskFilters
+  extractEnabledTaskFilters,
+  parseFilterOptions,
 } from '../util/index';
 
 class TaskTableTabUnconnected extends React.PureComponent {
@@ -54,21 +55,6 @@ class TaskTableTabUnconnected extends React.PureComponent {
     this.setState({ searchText: '', searchValue: '' });
   };
 
-  // TODO: Move this to the index file.
-  // This one is actually very generic and should work for any filter where the values match up
-  // With the database field values
-  parseIssueTypeFilterOptions = (issueTypeCounts) =>
-    Object.entries(issueTypeCounts).map(([key, issueTypeCount]) => {
-      let taskInfo;
-
-      taskInfo = {
-        value: key,
-        displayText: `${key} (${issueTypeCount})`
-      };
-
-      return { ...taskInfo, checked: false };
-    });
-
   getTableColumns = () => [
     claimantColumn(),
     {
@@ -84,7 +70,7 @@ class TaskTableTabUnconnected extends React.PureComponent {
     issueCountColumn(),
     {
       ...issueTypesColumn(),
-      filterOptions: this.parseIssueTypeFilterOptions(this.props.filterableTaskIssueTypes)
+      filterOptions: parseFilterOptions(this.props.filterableTaskIssueTypes)
     },
   ];
 
@@ -94,8 +80,6 @@ class TaskTableTabUnconnected extends React.PureComponent {
 
   render = () => {
     this.props.tabPaginationOptions[QUEUE_CONFIG.SEARCH_QUERY_REQUEST_PARAM] = this.state.searchValue;
-
-    console.log(this.props);
 
     return <React.Fragment>
       {this.props.description && <div className="cf-noncomp-queue-completed-task">{this.props.description}</div>}
